undo real->double compiler warning fixes

+++ src/mo_additional_geometry.f90
@@ -157,7 +157,9 @@ MODULE mo_additional_geometry
     TYPE(cartesian_coordinates)  :: normal_poly_edge !  normal vector
     TYPE(cartesian_coordinates)  :: normal_poly1_ips !  normal vector
     
-    REAL(KIND=wp) :: s_rot ! help variable for sense of rotation
+    REAL :: s_rot ! help variable for sense of rotation
+
+
 
     inflag = 0 ! set to a default value
     crossn = 0 ! start with crossn = 0
